requirements.txt
aiohttp==3.6.2
assertpy==1.1
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord==1.0.1
discord.py==1.3.4
idna==2.10
more-itertools==8.4.0
multidict==4.7.6
python-dotenv==0.14.0
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.5.0
